# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent API",
    "version": "1.0.0"
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/suites": {
      "get": {
        "operationId": "listSuites",
        "parameters": [],
        "tags": ["suites"],
        "description": "List all evaluation suites with run counts",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "completedCount": {
                        "type": "number"
                      },
                      "created_at": {
                        "type": "number"
                      },
                      "failedCount": {
                        "type": "number"
                      },
                      "id": {
                        "type": "string"
                      },
                      "lastStartedAt": {
                        "anyOf": [
                          {
                            "type": "null"
                          },
                          {
                            "type": "number"
                          }
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "runCount": {
                        "type": "number"
                      },
                      "runningCount": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "completedCount",
                      "created_at",
                      "failedCount",
                      "id",
                      "lastStartedAt",
                      "name",
                      "runCount",
                      "runningCount"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/compare/runs": {
      "get": {
        "operationId": "listCompletedRuns",
        "parameters": [],
        "tags": ["compare"],
        "description": "List completed runs available for comparison",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunRow"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RunRow": {
        "type": "object",
        "properties": {
          "config": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": {}
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "number"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "started_at": {
            "type": "number"
          },
          "status": {
            "anyOf": [
              {
                "enum": ["running"],
                "type": "string"
              },
              {
                "enum": ["completed"],
                "type": "string"
              },
              {
                "enum": ["failed"],
                "type": "string"
              }
            ]
          },
          "suite_id": {
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RunSummary"
              }
            ]
          }
        },
        "required": [
          "config",
          "finished_at",
          "id",
          "model",
          "name",
          "started_at",
          "status",
          "suite_id",
          "summary"
        ],
        "additionalProperties": false
      },
      "RunSummary": {
        "type": "object",
        "properties": {
          "failCount": {
            "type": "number"
          },
          "meanScores": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "passCount": {
            "type": "number"
          },
          "totalCases": {
            "type": "number"
          },
          "totalLatencyMs": {
            "type": "number"
          },
          "totalTokensIn": {
            "type": "number"
          },
          "totalTokensOut": {
            "type": "number"
          }
        },
        "required": [
          "failCount",
          "meanScores",
          "passCount",
          "totalCases",
          "totalLatencyMs",
          "totalTokensIn",
          "totalTokensOut"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use Authorization: Bearer <token> for session tokens or API keys."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Use X-API-Key: <key>."
      }
    }
  }
}